Quantification types in client

+++ datalayer/app/models/base_model.go
@@ -1,7 +1,7 @@
 package models
 
 type BaseDbModel struct {
-	Id uint `gorm:"primaryKey"`
+	Id uint64 `gorm:"primaryKey"`
 }
 
 type MonetaryAmount float64
